requirements.txt
python-instagram
scipy
matplotlib
stop-words
git+git://github.com/amueller/word_cloud.git